+++ lib/hoodoo/active/active_record/dated.rb
@@ -177,6 +177,17 @@ module Hoodoo
 
           self.nz_co_loyalty_hoodoo_dated_with = history_klass
 
+          # Enable the monkey patch to the Finder module's '#acquire_in' class
+          # method, if need be.
+
+          if self.include?( Hoodoo::ActiveRecord::Finder )
+            Hoodoo::Monkey.register(
+              target_unit:      self,
+              extension_module: Hoodoo::Monkey::Patch::ActiveRecordDatedFinderAdditions
+            )
+
+            Hoodoo::Monkey.enable( extension_module: Hoodoo::Monkey::Patch::ActiveRecordDatedFinderAdditions )
+          end
         end
 
         # If a prior call has been made to #dating_enabled then this method
@@ -289,7 +300,7 @@ module Hoodoo
         #
         # +unquoted_column_names+:: (Optional) An Array of Strings giving one
         #                           or more column names to use for the query.
-        #                           If omitted, all model attribtues are used
+        #                           If omitted, all model attributes are used
         #                           as columns. If the "id" column is not
         #                           included in the Array, it will be added
         #                           anyway as this column is mandatory. The