+++ lib/hoodoo/monkey.rb
@@ -10,6 +10,9 @@
 ########################################################################
 
 require 'hoodoo/monkey/monkey'
+
+require 'hoodoo/monkey/patch/active_record_dated_finder_additions'
+require 'hoodoo/monkey/patch/active_record_manually_dated_finder_additions'
+require 'hoodoo/monkey/patch/datadog_traced_amqp'
 require 'hoodoo/monkey/patch/newrelic_middleware_analytics'
 require 'hoodoo/monkey/patch/newrelic_traced_amqp'
-require 'hoodoo/monkey/patch/datadog_traced_amqp'