+++ spec/active/active_record/dated_spec.rb
@@ -5,44 +5,47 @@ describe Hoodoo::ActiveRecord::Dated do
 
   before :all do
     spec_helper_silence_stdout() do
-
-      ActiveRecord::Migration.create_table( :r_spec_model_effective_date_tests, :id => false ) do | t |
-        t.text :id,  :null => false
+      ActiveRecord::Migration.create_table( :r_spec_model_effective_date_tests, :id => :string ) do | t |
         t.text :data
         t.timestamps :null => true
       end
 
-      ActiveRecord::Migration.create_table( :r_spec_model_effective_date_tests_history_entries, :id => false ) do | t |
-        t.text     :id,              :null => false
+      ActiveRecord::Migration.change_column( :r_spec_model_effective_date_tests, :id, :string, :limit => 32 )
+
+      ActiveRecord::Migration.create_table( :r_spec_model_effective_date_tests_history_entries, :id => :string ) do | t |
         t.text     :uuid,            :null => false
         t.text     :data
         t.datetime :effective_start, :null => false
         t.datetime :effective_end,   :null => false
+
         t.timestamps :null => true
       end
 
       class RSpecModelEffectiveDateTest < ActiveRecord::Base
         include Hoodoo::ActiveRecord::Dated
+
         dating_enabled()
       end
 
-      ActiveRecord::Migration.create_table( :r_spec_model_effective_date_test_overrides, :id => false ) do | t |
-        t.text :id,  :null => false
+      ActiveRecord::Migration.create_table( :r_spec_model_effective_date_test_overrides, :id => :string ) do | t |
         t.text :data
         t.timestamps :null => true
       end
 
-      ActiveRecord::Migration.create_table( :r_spec_model_effective_date_history_entries, :id => false ) do | t |
-        t.text     :id,              :null => false
+      ActiveRecord::Migration.change_column( :r_spec_model_effective_date_test_overrides, :id, :string, :limit => 32 )
+
+      ActiveRecord::Migration.create_table( :r_spec_model_effective_date_history_entries, :id => :string ) do | t |
         t.text     :uuid,            :null => false
         t.text     :data
         t.datetime :effective_start, :null => false
         t.datetime :effective_end,   :null => false
+
         t.timestamps :null => true
       end
 
       class RSpecModelEffectiveDateTestOverride < ActiveRecord::Base
         include Hoodoo::ActiveRecord::Dated
+        include Hoodoo::ActiveRecord::Finder
 
         dating_enabled( :history_table_name => :r_spec_model_effective_date_history_entries )
       end
@@ -51,7 +54,16 @@ describe Hoodoo::ActiveRecord::Dated do
 
   shared_examples Hoodoo::ActiveRecord::Dated do
 
-    before( :all ) do
+    before( :each ) do
+      interaction = Hoodoo::Services::Middleware::Interaction.new( {}, nil )
+      interaction.context = Hoodoo::Services::Context.new(
+        Hoodoo::Services::Session.new,
+        interaction.context.request,
+        interaction.context.response,
+        interaction
+      )
+
+      @context = interaction.context
 
       # Create some example data for finding. The data has two different UUIDs
       # which I'll referer to as A and B. The following tables contain the
@@ -169,26 +181,16 @@ describe Hoodoo::ActiveRecord::Dated do
 
     context '#dated' do
       it 'returns counts correctly' do
-        # The contents of the Context are irrelevant aside from the fact that it
-        # needs a request to store the dated_at value.
-        request = Hoodoo::Services::Request.new
-        context = Hoodoo::Services::Context.new( nil, request, nil, nil )
+        @context.request.dated_at = @now - 10.hours
+        expect( model_klass.dated( @context ).count ).to be 0
 
-        context.request.dated_at = @now - 10.hours
-        expect( model_klass.dated( context ).count ).to be 0
-
-        context.request.dated_at = @now
-        expect( model_klass.dated( context ).count ).to be 2
+        @context.request.dated_at = @now
+        expect( model_klass.dated( @context ).count ).to be 2
       end
 
       def test_expectation( time, expected_data )
-        # The contents of the Context are irrelevant aside from the fact that it
-        # needs a request to store the dated_at value.
-        request = Hoodoo::Services::Request.new
-        context = Hoodoo::Services::Context.new( nil, request, nil, nil )
-        context.request.dated_at = time
-
-        expect( model_klass.dated( context ).pluck( :data ) ).to match_array( expected_data )
+        @context.request.dated_at = time
+        expect( model_klass.dated( @context ).pluck( :data ) ).to match_array( expected_data )
       end
 
       it 'returns no records before any were effective' do
@@ -208,27 +210,14 @@ describe Hoodoo::ActiveRecord::Dated do
       end
 
       it 'works with further filtering' do
-
-        # The contents of the Context are irrelevant aside from the fact that it
-        # needs a request to store the dated_at value.
-        request = Hoodoo::Services::Request.new
-        context = Hoodoo::Services::Context.new( nil, request, nil, nil )
-        context.request.dated_at = @now
-
-        expect( model_klass.dated( context ).where( :id => @uuid_a ).pluck( :data ) ).to eq( [ "six" ] )
+        @context.request.dated_at = @now
+        expect( model_klass.dated( @context ).where( :id => @uuid_a ).pluck( :data ) ).to eq( [ "six" ] )
       end
 
       it 'works with dating last' do
-
-        # The contents of the Context are irrelevant aside from the fact that it
-        # needs a request to store the dated_at value.
-        request = Hoodoo::Services::Request.new
-        context = Hoodoo::Services::Context.new( nil, request, nil, nil )
-        context.request.dated_at = @now
-
-        expect( model_klass.where( :id => @uuid_a ).dated( context ).pluck( :data ) ).to eq( [ "six" ] )
+        @context.request.dated_at = @now
+        expect( model_klass.where( :id => @uuid_a ).dated( @context ).pluck( :data ) ).to eq( [ "six" ] )
       end
-
     end
 
     context '#dated_historical_and_current' do
@@ -268,7 +257,171 @@ describe Hoodoo::ActiveRecord::Dated do
       end
     end
 
-  end
+    # These appear in the shared examples to take advantage of the test
+    # data that's set up, but only work when "model_klass" is the
+    # Finder-extended "RSpecModelEffectiveDateTestOverride".
+    #
+    # Hence "next unless..." in each example (using "skip" in a "before"
+    # filter would result in RSpect listing pending tests; they are not
+    # pending at all).
+
+    context 'with Hoodoo::ActiveRecord::Finder support' do
+      context '#scoped_in' do
+        it 'generates appropriate scope' do
+          next unless model_klass == RSpecModelEffectiveDateTestOverride
+
+          expect( Hoodoo::ActiveRecord::Support ).to(
+            receive( :full_scope_for ).once().with(
+              RSpecModelEffectiveDateTestOverride, @context
+            ).and_call_original()
+          )
+
+          sql = RSpecModelEffectiveDateTestOverride.scoped_in( @context ).to_sql
+
+          expect( sql ).to include( "UNION ALL" )
+          expect( sql ).to include( "' AND (\"effective_end\" > '" )
+        end
+
+        it 'generates appropriate undated scope' do
+          next unless model_klass == RSpecModelEffectiveDateTestOverride
+
+          expect( Hoodoo::ActiveRecord::Support ).to(
+            receive( :add_undated_scope_to ).once().with(
+              kind_of( ActiveRecord::Relation ), RSpecModelEffectiveDateTestOverride, @context
+            ).and_call_original()
+          )
+
+          sql = RSpecModelEffectiveDateTestOverride.scoped_undated_in( @context ).to_sql
+
+          expect( sql ).to eq( "SELECT \"r_spec_model_effective_date_test_overrides\".* FROM \"r_spec_model_effective_date_test_overrides\"" )
+        end
+      end
+
+      context '#acquire_in' do
+        context 'with contemporary' do
+          it 'finds the contemporary record' do
+            next unless model_klass == RSpecModelEffectiveDateTestOverride
+
+            @context.request.dated_at            = nil
+            @context.request.uri_path_components = [ @uuid_a ]
+
+            result = RSpecModelEffectiveDateTestOverride.acquire_in_and_update( @context )
+            expect( result ).to_not be_nil
+
+            expect( @context.response.halt_processing? ).to eq( false )
+            expect( result.data                        ).to eq( 'six' )
+          end
+
+          it 'finds a historic record' do
+            next unless model_klass == RSpecModelEffectiveDateTestOverride
+
+            @context.request.dated_at            = @now - 4.hours
+            @context.request.uri_path_components = [ @uuid_a ]
+
+            result = RSpecModelEffectiveDateTestOverride.acquire_in_and_update( @context )
+            expect( result ).to_not be_nil
+
+            expect( @context.response.halt_processing? ).to eq( false )
+            expect( result.data                        ).to eq( 'one' )
+          end
+
+          it 'indicates correctly that a contemporary exists during a far-backdated lookup' do
+            next unless model_klass == RSpecModelEffectiveDateTestOverride
+
+            @context.request.dated_at            = @now - 1.year
+            @context.request.uri_path_components = [ @uuid_a ]
+
+            result = RSpecModelEffectiveDateTestOverride.acquire_in_and_update( @context )
+            expect( result ).to be_nil
+
+            expect( @context.response.halt_processing?    ).to eq( true )
+            expect( @context.response.errors.errors.count ).to eq( 2    )
+
+            expect( @context.response.errors.errors[ 0 ][ 'code'      ] ).to eq( 'generic.not_found' )
+            expect( @context.response.errors.errors[ 1 ][ 'code'      ] ).to eq( 'generic.contemporary_exists' )
+            expect( @context.response.errors.errors[ 0 ][ 'reference' ] ).to eq( @uuid_a )
+            expect( @context.response.errors.errors[ 1 ][ 'reference' ] ).to eq( @uuid_a )
+          end
+        end
+
+        context 'with a UUID that matches no existing record' do
+          it 'indicates correctly that no record exists and does not say a contemporary is present' do
+            next unless model_klass == RSpecModelEffectiveDateTestOverride
+
+            alt_uuid                             = Hoodoo::UUID.generate
+            @context.request.dated_at            = @now - 5.seconds
+            @context.request.uri_path_components = [ alt_uuid ]
+
+            result = RSpecModelEffectiveDateTestOverride.acquire_in_and_update( @context )
+            expect( result ).to be_nil
+
+            expect( @context.response.halt_processing?    ).to eq( true )
+            expect( @context.response.errors.errors.count ).to eq( 1    )
+
+            expect( @context.response.errors.errors[ 0 ][ 'code'      ] ).to eq( 'generic.not_found' )
+            expect( @context.response.errors.errors[ 0 ][ 'reference' ] ).to eq( alt_uuid )
+          end
+        end
+
+        context 'without contemporary' do
+          before( :each ) do
+            if model_klass == RSpecModelEffectiveDateTestOverride
+              record = RSpecModelEffectiveDateTestOverride.find( @uuid_a )
+              record.destroy!
+
+              sleep( 0.1 ) # Be sure that there's nothing left at "now", within timer resolution
+            end
+          end
+
+          it 'finds no contemporary record' do
+            next unless model_klass == RSpecModelEffectiveDateTestOverride
+
+            @context.request.dated_at            = nil
+            @context.request.uri_path_components = [ @uuid_a ]
+
+            result = RSpecModelEffectiveDateTestOverride.acquire_in_and_update( @context )
+            expect( result ).to be_nil
+
+            expect( @context.response.halt_processing?    ).to eq( true )
+            expect( @context.response.errors.errors.count ).to eq( 1    )
+
+            expect( @context.response.errors.errors[ 0 ][ 'code'      ] ).to eq( 'generic.not_found' )
+            expect( @context.response.errors.errors[ 0 ][ 'reference' ] ).to eq( @uuid_a )
+          end
+
+          it 'finds a historic record' do
+            next unless model_klass == RSpecModelEffectiveDateTestOverride
+
+            @context.request.dated_at            = @now - 4.hours
+            @context.request.uri_path_components = [ @uuid_a ]
+
+            result = RSpecModelEffectiveDateTestOverride.acquire_in_and_update( @context )
+            expect( result ).to_not be_nil
+
+            expect( @context.response.halt_processing? ).to eq( false )
+            expect( result.data                        ).to eq( 'one' )
+          end
+
+          it 'indicates correctly that no contemporary exists during a far-backdated lookup' do
+            next unless model_klass == RSpecModelEffectiveDateTestOverride
+
+            @context.request.dated_at            = @now - 1.year
+            @context.request.uri_path_components = [ @uuid_a ]
+
+            result = RSpecModelEffectiveDateTestOverride.acquire_in_and_update( @context )
+            expect( result ).to be_nil
+
+            expect( @context.response.halt_processing?    ).to eq( true )
+            expect( @context.response.errors.errors.count ).to eq( 1    )
+
+            expect( @context.response.errors.errors[ 0 ][ 'code'      ] ).to eq( 'generic.not_found' )
+            expect( @context.response.errors.errors[ 0 ][ 'reference' ] ).to eq( @uuid_a )
+          end
+        end
+      end
+    end
+
+  end # 'shared_examples Hoodoo::ActiveRecord::Dated do'
 
   context "using default effective dating config" do